requirements.txt
Django
psycopg2
pandas
regex-generator-lib